requirements.txt
flask[async]
Flask>=2.0
aiohttp>=3.7
python-dotenv>=0.19
hypercorn>=0.13
# openai

